refactor: unify plugin callback — single onChartClick, shared PluginProps
- New: plugins/utils.ts — PluginProps type + useEChartsClick() hook
- All 16 plugins use PluginProps (no more per-plugin interfaces)
- chart-renderer passes only onChartClick (removed handleEChartsClick)
- ECharts plugins wrap onChartClick via useEChartsClick() internally
- Removed unused useMemo, EChartsClickEvent imports from renderer
